# back/openapi.yaml
openapi: 3.0.3
info:
  title: ERPlay Public API
  version: 1.2.0
  description: |
    API pública utilizada por las aplicaciones web de alumnos y supervisores de ERPlay.
    Todos los endpoints requieren autenticación mediante JWT Bearer salvo los de autenticación.
servers:
  - url: https://api.erplay.local/api
    description: Entorno de producción (ejemplo)
  - url: http://localhost:3000/api
    description: Entorno de desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /questions:
    post:
      tags: [Questions]
      summary: Crear una pregunta
      operationId: createQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionCreateRequest"
      responses:
        "201":
          description: Pregunta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionCreatedResponse"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /questions/{questionId}/verify:
    parameters:
      - $ref: "#/components/parameters/QuestionIdParam"
    post:
      tags: [Questions]
      summary: Verificar una pregunta
      operationId: verifyQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionVerifyRequest"
      responses:
        "200":
          description: Revisión aplicada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: No autorizado para revisar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Pregunta no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    QuestionIdParam:
      name: questionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Identificador de la pregunta
...
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      additionalProperties: false
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
      additionalProperties: true
...
    QuestionCreateRequest:
      type: object
      required: [diagramId, prompt, hint, options, correctIndex]
      properties:
        diagramId:
          type: string
          format: uuid
        prompt:
          type: string
        hint:
          type: string
        options:
          type: array
          minItems: 2
          items:
            type: string
        correctIndex:
          type: integer
          minimum: 0
      additionalProperties: false
    QuestionCreatedResponse:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
...
    QuestionVerifyRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approve, reject]
        comment:
          type: string
          nullable: true
      additionalProperties: false
...